… src/Http/Http.Abstractions/src/Internal/PathStringHelper.cs