- Replace `PathStringHelper.IsValidPathChar` loops with `SearchValues`
- Add a vectorized fast-path for longer sections of valid chars
- Use existing `Uri.IsHexEncoding` helper instead of `IsPercentEncodedChar`
- Remove some dead code and eliminate some redundant length checks

Contributes to #46484

I'll post some benchmark numbers later today, but you can easily expect a 10x improvement for long already-valid values.