@MihaZupan 2023-06-30
[Benchmark code](https://gist.github.com/MihaZupan/0b5561bcc4e40e054fd6f9be58b077b8)

| Method |                 Path |       Mean |     Error |    StdDev | Ratio |
|------- |--------------------- |-----------:|----------:|----------:|------:|
| Before |         LongTestPath | 283.907 ns | 2.8201 ns | 4.2210 ns | 16.37 |
|  After |         LongTestPath |  17.331 ns | 0.0495 ns | 0.0693 ns |  1.00 |
|        |                      |            |           |           |       |
| Before | LongT(...)rcent [24] | 302.893 ns | 0.7137 ns | 1.0683 ns | 11.90 |
|  After | LongT(...)rcent [24] |  25.450 ns | 0.0900 ns | 0.1291 ns |  1.00 |
|        |                      |            |           |           |       |
| Before |    LongValidTestPath | 244.094 ns | 1.3209 ns | 1.9771 ns | 32.45 |
|  After |    LongValidTestPath |   7.523 ns | 0.0330 ns | 0.0483 ns |  1.00 |
|        |                      |            |           |           |       |
| Before |             TestPath |  24.910 ns | 0.2281 ns | 0.3271 ns |  1.83 |
|  After |             TestPath |  13.643 ns | 0.0851 ns | 0.1193 ns |  1.00 |