@Tratcher 2023-06-30
components-e2e is being addressed: https://github.com/dotnet/aspnetcore/pull/49113