@mitchdenny 2023-07-03
Also related: https://github.com/dotnet/aspnetcore/pull/49114